+++ tests/helpers/sane.js
@@ -1,7 +1,10 @@
-var path             = require('path');
-var root             = process.cwd();
+'use strict';
 
-var saneExec   = (process.platform === 'win32' ? 'sane.cmd' : 'sane');
-var sane       = path.join(root, 'bin', saneExec);
+var path = require('path');
+
+
+var root      = process.cwd();
+var saneExec  = (process.platform === 'win32' ? 'sane.cmd' : 'sane');
+var sane      = path.join(root, 'bin', saneExec);
 
 module.exports = sane;