+++ tests/lint/eslintTest.js
@@ -1,0 +1,15 @@
+/*eslint-env node, mocha, es6 */
+'use strict';
+
+var lint = require('mocha-eslint');
+
+
+var paths = [
+  'bin/sane',
+  'lib',
+  'tests'
+];
+var options = {};
+options.formatter = 'stylish';
+
+lint(paths, options);