+++ tests/runner.js
@@ -1,7 +1,9 @@
+/*eslint no-process-exit:0*/
 'use strict';
 
-var glob = require('glob');
+var glob  = require('glob');
 var Mocha = require('mocha');
+var chalk = require('chalk');
 
 
 require('traceur').require.makeDefault(function (filename) {
@@ -17,8 +19,15 @@ var mocha = new Mocha({
   reporter: 'spec'
 });
 
-// var arg = process.argv[2];
-var root = 'tests/{unit,acceptance}';
+// Determine which tests to run based on argument passed to runner
+var arg = process.argv[2];
+if (!arg) {
+  var root = 'tests/{unit,acceptance,lint}';
+} else if (arg === 'lint') {
+  var root = 'tests/lint';
+} else {
+  var root = 'tests/{unit,acceptance}';
+}
 
 function addFiles(mocha, files) {
   glob.sync(root + files).forEach(mocha.addFile.bind(mocha));
@@ -26,12 +35,15 @@ function addFiles(mocha, files) {
 
 addFiles(mocha, '/**/*Test.js');
 
-// if (arg === 'all') {
-//   addFiles(mocha, '/**/*-slow.js');
-// }
-
 mocha.run(function (failures) {
   process.on('exit', function () {
+    if (failures === 1) {
+      console.log(chalk.red('1 Failing Test'));
+    } else if (failures > 1) {
+      console.log(chalk.red(failures, 'Failing Tests'));
+    } else if (failures === 0) {
+      console.log(chalk.green('All Tests Passed!'));
+    }
     process.exit(failures);
   });
 });