+++ tests/unit/jshintTest.js
@@ -1,8 +1,8 @@
-'use strict';
+// 'use strict';
 
-require('mocha-jshint')([
-// Both use async/await, currently unsupported by jshint or eslint 
-// 'tests',
-// 'lib',
-'bin',
-]);
+// require('mocha-jshint')([
+// // Both use async/await, currently unsupported by jshint or eslint
+// // 'tests',
+// // 'lib',
+// 'bin',
+// ]);